Fix LoadError: extract Compliance::Countries to its own file for Zeitwerk autoloading
Sentry reports `LoadError: cannot load such file -- compliance/countries`
originating from `Kernel#require`.

Root cause: `Compliance::Countries` is defined inline in
`lib/utilities/compliance.rb`, but Rails uses Zeitwerk autoloading
(`config.load_defaults 7.0`) with `lib/utilities` in `eager_load_paths`.
Zeitwerk expects `Compliance::Countries` to live in
`lib/utilities/compliance/countries.rb`. When something references
`Compliance::Countries` before `compliance.rb` happens to be loaded,
Zeitwerk tries to autoload from `compliance/countries.rb`, which
doesn't exist, and raises LoadError.

Fix: extract the `Compliance::Countries` module into
`lib/utilities/compliance/countries.rb` so its file path matches
Zeitwerk's expected layout. The remaining top-level `Compliance`
constants (`DEFAULT_TOS_VIOLATION_REASON`,
`EXPLICIT_NSFW_TOS_VIOLATION_REASON`, `TOS_VIOLATION_REASONS`,
`VAT_EXEMPT_REGIONS`) stay in `lib/utilities/compliance.rb`.

No behavior change — module contents are moved verbatim.
